Refactor multi-src theme name
